ENH: enhanced compatibility with ImageFunctionAdapter

+++ Code/FeatureExtraction/otbFlusserMomentsImageFunction.h
@@ -58,7 +58,7 @@ namespace otb
  * \ingroup ImageFunctions
  */
 
-template <class TInputImage, class TCoordRep = float >
+template <class TInputImage, class TCoordRep = double >
 class ITK_EXPORT FlusserMomentsImageFunction :
   public itk::ImageFunction< TInputImage,
     itk::FixedArray<
