+++ Code/FeatureExtraction/otbRadiometricMomentsImageFunction.h
@@ -37,7 +37,7 @@ namespace otb
  */
 
 
-template <class TInputImage, class TCoordRep = float >
+template <class TInputImage, class TCoordRep = double >
 class ITK_EXPORT RadiometricMomentsImageFunction :
   public itk::ImageFunction< TInputImage,
     itk::FixedArray<
